Use kebabdropdown for the settings menu
Re-use Cockpit's Kebab dropdown component.
